[Chore 159910058] Rename app.js to reflect the content, which versioning for api
